+++ modules/Content/cli.php
@@ -1,0 +1,8 @@
+<?php
+
+if (!isset($cli, $app) || PHP_SAPI !== 'cli') {
+    return;
+}
+
+$cli->add(new Content\Command\Field\Remove($app));
+$cli->add(new Content\Command\Field\Rename($app));