+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix initial item position on tree item creation
 - Add avif image type support (depending on used GD lib)
+- Add experimental natural language to mongo query support (only for admin ui filtering collection items and assets)
 
 ## 2.4.1 (2023-03-20)