+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix initial item position on tree item creation
 - Add avif image type support (depending on used GD lib)
 - Add experimental natural language to mongo query support (only for admin ui filtering collection items and assets)
+- Add lightweight indexing library (for full-text search)
 
 ## 2.4.1 (2023-03-20)