+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add avif image type support (depending on used GD lib)
 - Add experimental natural language to mongo query support (only for admin ui filtering collection items and assets)
 - Add lightweight indexing library (for full-text search)
+- Make logged login data configurable
 
 ## 2.4.1 (2023-03-20)