+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add experimental natural language to mongo query support (only for admin ui filtering collection items and assets)
 - Add lightweight indexing library (for full-text search)
 - Make logged login data configurable
+- Add filter models feature on role permission page
 
 ## 2.4.1 (2023-03-20)