+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add lightweight indexing library (for full-text search)
 - Make logged login data configurable
 - Add filter models feature on role permission page
+- Fix populate option singleton data queries vi GraphQL
 
 ## 2.4.1 (2023-03-20)